Update User from Adminpanel

See merge request informatik/qe/swess23/group2/g2t4!19

+++ backend/src/main/java/at/qe/backend/services/UserxService.java
@@ -1,13 +1,11 @@
 package at.qe.backend.services;
 
-import at.qe.backend.models.Userx;
 import at.qe.backend.models.UserRole;
+import at.qe.backend.models.Userx;
 import at.qe.backend.repositories.UserxRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 import java.util.Collection;
@@ -51,20 +49,20 @@ public class UserxService {
      * Saves the user. This method will also set {@link Userx#createDate} for new
      * entities or {@link Userx#updateDate} for updated entities. The user
      * requesting this operation will also be stored as {@link Userx#createDate}
-     * or {@link Userx#updateUser} respectively.
+     * or {@link Userx#updateUserUsername} respectively.
      *
      * @param user the user to save
      * @return the updated user
      */
     @PreAuthorize("hasAuthority('ADMIN')")
-    public Userx saveUser(Userx user) {
+    public Userx saveUser(Userx user, String username_session) {
         if (user.isNew()) {
             user.setCreateDate(new Date());
-            user.setCreateUser(getAuthenticatedUser());
+            user.setCreateUserUsername(username_session);
             user.setRoles(Set.of(UserRole.USER));
         } else {
             user.setUpdateDate(new Date());
-            user.setUpdateUser(getAuthenticatedUser());
+            user.setUpdateUserUsername(username_session);
         }
         return userRepository.save(user);
     }
@@ -75,17 +73,13 @@ public class UserxService {
      * @param user the user to delete
      */
     @PreAuthorize("hasAuthority('ADMIN')")
-    public void deleteUser(Userx user) {
+    public boolean deleteUser(Userx user) {
+        if(user.getRoles().contains(UserRole.ADMIN) && userRepository.countUserxByRolesContaining(UserRole.ADMIN)<=1){
+            //TODO Add log event here:
+            return false;
+        }
         userRepository.delete(user);
-    }
-
-
-    /**
-     * @return user that is currently signed in
-     */
-    private Userx getAuthenticatedUser() {
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        return userRepository.findFirstByUsername(auth.getName());
+        return true;
     }
 }
